+++ packages/big-design/src/components/InlineAlert/index.ts
@@ -1,4 +1,0 @@
-import { InlineAlertProps as _InlineAlertProps } from './InlineAlert';
-
-export { InlineAlert } from './InlineAlert';
-export type InlineAlertProps = _InlineAlertProps;