+++ packages/big-design/src/components/InlineMessage/index.ts
@@ -1,0 +1,4 @@
+import { InlineMessageProps as _InlineMessageProps } from './InlineMessage';
+
+export { InlineMessage } from './InlineMessage';
+export type InlineMessageProps = _InlineMessageProps;